moving room client logic to react context
